# out/src/Johnfg2610.Stockit/wwwroot/openapi-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "contact" : {
      "email" : "johnfg2610@gmail.com",
      "name" : "john",
      "url" : "https://www.github.com/johnfg2610/stockit"
    },
    "description" : "Used to access inventory counts as well as brands and catalog items",
    "license" : {
      "name" : "ISC",
      "url" : "https://github.com/johnfg2610/stockit/blob/master/LICENSE"
    },
    "title" : "Inventory",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "http://localhost:3000"
  } ],
  "security" : [ {
    "jwt-auth" : [ ]
  } ],
...
  "paths" : {
...
    "/catalog" : {
      "post" : {
        "description" : "Create a new catalog item",
        "operationId" : "post-catalog",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CatalogItemForm"
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "Created"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Bad Request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Not Found"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Create a new catalog",
        "tags" : [ "catalog" ]
      }
    },
    "/catalog/{id}" : {
      "delete" : {
        "description" : "Delete catalog by ID",
        "operationId" : "delete-catalog-id",
        "parameters" : [ {
          "description" : "The identifier to refrence the catalog item",
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "201" : {
            "description" : "Created"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Bad Request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Not Found"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Delete catalog by ID",
        "tags" : [ "catalog" ]
      },
      "get" : {
        "description" : "Gets a catalog item by ID",
        "operationId" : "get-catalog-id",
        "parameters" : [ {
          "description" : "The identifier to refrence the catalog item",
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CatalogItem"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Bad Request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Not Found"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Get catalog by ID",
        "tags" : [ "catalog" ]
      },
      "put" : {
        "description" : "Replaces the catalog item",
        "operationId" : "put-catalog-id",
        "parameters" : [ {
          "description" : "The identifier to refrence the catalog item",
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CatalogItemForm"
              }
            },
            "application/xml" : {
              "schema" : {
                "properties" : { },
                "type" : "object"
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "Created"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Bad Request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Not Found"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Replace catalog by ID",
        "tags" : [ "catalog" ]
      }
    }
  },
  "components" : {
...
    "schemas" : {
      "CatalogItem" : {
        "description" : "A item contained within the catalog",
        "example" : {
          "imagesUrls" : [ "https://openapi-generator.tech", "https://openapi-generator.tech" ],
          "sellerId" : "sellerId",
          "_links" : {
            "seller" : "https://openapi-generator.tech",
            "brand" : "https://openapi-generator.tech"
          },
          "brandId" : "brandId",
          "name" : "name",
          "description" : "description",
          "id" : "id",
          "iconUrl" : "https://openapi-generator.tech"
        },
        "properties" : {
          "id" : {
            "description" : "A random identifier used to identify this object",
            "type" : "string"
          },
          "name" : {
            "description" : "the name of this item",
            "type" : "string"
          },
          "description" : {
            "description" : "the description of this item",
            "type" : "string"
          },
          "iconUrl" : {
            "description" : "A optional url pointing to a jpg or png image",
            "format" : "uri",
            "type" : "string"
          },
          "imagesUrls" : {
            "description" : "A optional list of image urls to display",
            "items" : {
              "format" : "uri",
              "type" : "string"
            },
            "type" : "array"
          },
          "_links" : {
            "$ref" : "#/components/schemas/CatalogItem__links"
          },
          "sellerId" : {
            "description" : "A random string used to identify the seller",
            "type" : "string"
          },
          "brandId" : {
            "description" : "A random string used to identify the brand",
            "type" : "string"
          }
        },
        "required" : [ "_links", "brandId", "description", "name", "sellerId" ],
        "title" : "CatalogItem",
        "type" : "object",
        "x-tags" : [ "catalog" ]
      },
      "CatalogItemForm" : {
        "description" : "A item contained within the catalog",
        "example" : {
          "imagesUrls" : [ "https://openapi-generator.tech", "https://openapi-generator.tech" ],
          "sellerId" : "sellerId",
          "_links" : {
            "seller" : "https://openapi-generator.tech",
            "brand" : "https://openapi-generator.tech"
          },
          "brandId" : "brandId",
          "name" : "name",
          "description" : "description",
          "iconUrl" : "https://openapi-generator.tech"
        },
        "properties" : {
          "name" : {
            "description" : "the name of this item",
            "type" : "string"
          },
          "description" : {
            "description" : "the description of this item",
            "type" : "string"
          },
          "iconUrl" : {
            "description" : "A optional url pointing to a jpg or png image",
            "format" : "uri",
            "type" : "string"
          },
          "imagesUrls" : {
            "description" : "A optional list of image urls to display",
            "items" : {
              "format" : "uri",
              "type" : "string"
            },
            "type" : "array"
          },
          "_links" : {
            "$ref" : "#/components/schemas/CatalogItem__links"
          },
          "sellerId" : {
            "description" : "A random string used to identify the seller",
            "type" : "string"
          },
          "brandId" : {
            "description" : "A random string used to identify the brand",
            "type" : "string"
          }
        },
        "required" : [ "_links", "brandId", "description", "name", "sellerId" ],
        "title" : "CatalogItemForm",
        "type" : "object",
        "x-tags" : [ "catalog" ]
      },
...
      "Error" : {
        "description" : "Used to describe a error that happened during the execution of a service",
        "properties" : {
          "code" : {
            "type" : "number"
          },
          "message" : {
            "type" : "string"
          },
          "requestId" : {
            "type" : "string"
          }
        },
        "title" : "Error",
        "type" : "object"
      },
...
      "CatalogItem__links" : {
        "description" : "A object containing links to other objects refrenced in this item",
        "example" : {
          "seller" : "https://openapi-generator.tech",
          "brand" : "https://openapi-generator.tech"
        },
        "properties" : {
          "brand" : {
            "description" : "a url to the brand who makes this product",
            "format" : "uri",
            "type" : "string"
          },
          "seller" : {
            "description" : "a url to the seller who sales this item",
            "format" : "uri",
            "type" : "string"
          }
        },
        "required" : [ "brand", "seller" ]
      }
    },
    "securitySchemes" : {
      "jwt-auth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  }
}